changed long int fields to long long and the functions that use stol to stoll

+++ part_1/askisi_2.cpp
@@ -13,8 +13,8 @@ typedef struct record
     string commodity;
     string transport_mode;
     string measure;
-    long int value;
-    long int cumulative;
+    long long value;
+    long long cumulative;
 } Record;
 
 
@@ -116,7 +116,7 @@ void heapSortCumulatives(vector<Record> &data, int size)
 
 void quickSortCumulatives(vector<Record> &data, int left, int right)
 {
-    long int pivot;
+    long long pivot;
     int leftArrow, rightArrow;
     leftArrow = left;
     rightArrow = right;
@@ -231,10 +231,10 @@ void insertDataToField(int fieldCounter, string token, Record& record)
         record.measure = token;
         break;
     case 8:
-        record.value = stol(token);
+        record.value = stoll(token);
         break;
     case 9:
-        record.cumulative = stol(token);
+        record.cumulative = stoll(token);
         break;
     default:
         break;
